apps/api/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
backoff==2.2.1
banks==2.2.0
beautifulsoup4==4.14.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
croniter==6.0.0
cryptography==46.0.2
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
emoji==2.15.0
fastapi==0.118.0
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.9.0
greenlet==3.2.4
griffe==1.14.0
grpcio==1.75.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
langdetect==1.0.9
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.4
llama-index-cli==0.5.3
llama-index-core==0.14.4
llama-index-embeddings-openai==0.5.1
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.1
llama-index-llms-openai==0.6.1
llama-index-readers-file==0.5.4
llama-index-readers-llama-parse==0.5.1
llama-index-workflows==2.6.0
llama-parse==0.6.54
lxml==6.0.2
MarkupSafe==3.0.3
marshmallow==3.26.1
multidict==6.6.4
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.2
numpy==2.3.3
olefile==0.47
openai==1.109.1
packaging==25.0
pandas==2.2.3
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
portalocker==3.2.0
propcache==0.4.0
protobuf==6.32.1
psutil==7.1.0
pycparser==2.23
pydantic==2.11.10
pydantic_core==2.33.2
Pygments==2.19.2
pypdf==6.1.1
pdfplumber==0.11.0
python-docx==1.1.2
pytest==8.4.2
pytest-asyncio==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.9
python-oxmsg==0.0.2
pytz==2025.2
# pywin32==311  # Windows-only package, not needed in Docker container
PyYAML==6.0.3
qdrant-client==1.15.1
RapidFuzz==3.14.1
redis==6.4.0
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rq==2.6.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
starlette==0.48.0
striprtf==0.0.26
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
unstructured==0.18.15
unstructured-client==0.42.3
urllib3==2.5.0
uvicorn==0.37.0
watchfiles==1.1.0
webencodings==0.5.1
websockets==15.0.1
wrapt==1.17.3
yarl==1.20.1

# PostgreSQL dependencies
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
